launch003 - adds tweaks to control loop
launch002 worked but there was a lot of vibration/oscillation in the launcher whilst the engine was firing, with the nozzle direction cycling up and down. The Kos control loop parameters can be tweaked and making the times longer seems to have damped out this problem, without any loss of control. I like my launchers to fly gently!