launch004 - tweaks for a different launch vehicle
This is basically launch003 with a few small changes. I wanted to see if it would work with a different (bigger!) launcher, and it does. The tweaks were largely experimental, e.g. checking whether it is best to launch with or without the first stage engine started up.

It works though and has been used on real missions now.